Bug fix, status endpoint impl.

+++ src/main/java/chat/tamtam/bot/domain/BotSchemeEntity.java
@@ -24,7 +24,7 @@ public class BotSchemeEntity {
     private Integer id;
 
     @Column(name = "userId")
-    private @NonNull Integer userId;
+    private @NonNull Long userId;
 
     @Column(name = "botId")
     private Long botId;
