+++ src/test/java/chat/tamtam/bot/TestContext.java
@@ -5,7 +5,7 @@ import chat.tamtam.bot.domain.BotSchemeEntity;
 public class TestContext {
 
     public static final int BOT_ID = 1;
-    public static final Integer USER_ID = 11;
+    public static final Long USER_ID = 11L;
     public static final int FAILED_BOT_ID = 2;
     public static final String BOT_NAME = "BOT NAME";
     public static final String AUTH_TOKEN = "AUTH";
